Move useTrigger hook to dedicated hooks directory
